add ArrayParser.kt

+++ src/parser/structure/CallChain.kt
@@ -1,6 +1,8 @@
 package parser.structure
 
+import java.math.BigInteger
+
 interface CallChain {
-    fun evaluate(arr: List<Int>): List<Int>
+    fun evaluate(arr: List<BigInteger>): List<BigInteger>
     override fun toString(): String
 }
