+++ src/parser/structure/callchain/CallChainImpl.kt
@@ -1,9 +1,10 @@
 package parser.structure.callchain
 
 import parser.structure.CallChain
+import java.math.BigInteger
 
 class CallChainImpl(private val call: Call, private val callChain: CallChain) : CallChain {
-    override fun evaluate(arr: List<Int>): List<Int> {
+    override fun evaluate(arr: List<BigInteger>): List<BigInteger> {
         return callChain.evaluate(call.evaluate(arr))
     }
